rename: server handlers 분리
